requirements.txt
absl-py==0.8.1
asgiref==3.2.8
# astor==0.8.0
astunparse==1.6.3
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.10
django-dotenv==1.4.2
gast==0.3.3
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.24.3
gunicorn==19.9.0
h5py==2.10.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
nltk==3.4.5
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.1.0
protobuf==3.10.0
psycopg2-binary==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
PySocks==1.7.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
scipy==1.4.1
six==1.12.0
sqlparse==0.3.0
tb-nightly==1.14.0a20190301
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.0.0
tensorflow-estimator==2.2.0
tensorflow-hub==0.7.0
termcolor==1.1.0
textblob==0.15.3
tf-estimator-nightly==1.14.0.dev2019030115
tweepy==3.8.0
tweet-preprocessor==0.5.0
urllib3==1.25.6
Werkzeug==0.16.0
whitenoise==4.1.4
wrapt==1.11.2
